Proper icons and menu items added
