--- src/app/my-nodes/layout.tsx
@@ -1,12 +1,0 @@
-export const metadata = {
-	title: 'Next.js',
-	description: 'Generated by Next.js'
-}
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-	return (
-		<html lang="en">
-			<body>{children}</body>
-		</html>
-	)
-}
