+++ src/components/ui/side-nav.tsx
@@ -1,9 +1,7 @@
 'use client'
 
-import { LogoIcon, MenuIcon } from '@/components/ui/icons'
-import Popover from '@/components/ui/popover'
 import Link from 'next/link'
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 
 interface MenuItem {
 	label: string
@@ -14,132 +12,94 @@ interface MenuItem {
 interface SideNavProps {
 	menuItems: MenuItem[]
 	menuItemsFooter: MenuItem[]
+	isExpanded: boolean
+	toggleSidebar: () => void
 }
 
-export default function SideNav({ menuItems, menuItemsFooter }: SideNavProps) {
-	const [hidden, setHidden] = useState(false)
-	const [active, setActive] = useState('Home')
-	const [isMobile, setIsMobile] = useState(false)
-	const [open, setOpen] = useState(false)
-
-	// Handle responsive behavior
-	useEffect(() => {
-		const handleResize = () => {
-			const width = window.innerWidth
-
-			if (width <= 430) {
-				setHidden(true) // Hide sidebar initially on iPhone
-				setIsMobile(true)
-			} else if (width <= 1024) {
-				setHidden(false)
-				setOpen(false) // Collapse sidebar on iPad
-				setIsMobile(false)
-			} else {
-				setHidden(false)
-				setOpen(true) // Expand sidebar on desktop
-				setIsMobile(false)
-			}
-		}
-
-		handleResize() // Set initial state
-		window.addEventListener('resize', handleResize)
-
-		return () => window.removeEventListener('resize', handleResize)
-	}, [])
-
-	// Toggle sidebar on mobile (iPhone only)
-	const toggleMobileNav = () => {
-		setHidden(!hidden)
-	}
-
+const SideNav: React.FC<SideNavProps> = ({
+	menuItems,
+	menuItemsFooter,
+	isExpanded,
+	toggleSidebar
+}) => {
 	return (
-		<section className="">
-			{/* Mobile Header (Only on iPhone) */}
-			{isMobile && (
-				<header className="fixed left-0 right-0 top-0 z-50 flex items-center justify-between bg-[#F0F0F0] px-4 py-3 shadow-md">
-					<LogoIcon />
-					<MenuIcon
-						className="cursor-pointer text-[#A0A2A0] duration-100 hover:text-black"
-						onClick={toggleMobileNav}
-					/>
-				</header>
-			)}
-
-			{/* Background Overlay (iPhone Only) */}
-			{isMobile && !hidden && (
+		<nav
+			id="sidenav"
+			className={`min-h-screen bg-[#F0F0F0] transition-all duration-300 ${
+				isExpanded ? 'w-[260px]' : 'w-[72px]'
+			}`}
+		>
+			{/* Sidebar Header */}
+			<div
+				id="sidebar_header"
+				className="bg-gray-300 flex h-[74px] items-center px-4 py-3 shadow-md"
+			>
+				{/* Logo - Hidden when collapsed */}
 				<div
-					className="fixed inset-0 z-40 bg-black/50"
-					onClick={toggleMobileNav} // Close sidebar when clicking outside
-				></div>
-			)}
-
-			{/* Sidebar */}
-			{!hidden && (
-				<nav
-					className={`fixed left-0 top-0 flex h-screen flex-col bg-[#F0F0F0] px-4 text-black shadow-md duration-300 ${
-						isMobile ? 'z-50 w-full' : open ? 'w-[260px]' : 'w-[72px]'
+					className={`transition-opacity duration-300 ${
+						isExpanded ? 'opacity-100' : 'hidden opacity-0'
 					}`}
 				>
-					{/* Header */}
-					<div className="flex h-14 items-center justify-between px-3 py-2">
-						{open && <LogoIcon />}
-						<MenuIcon
-							className={`cursor-pointer text-[#A0A2A0] duration-100 hover:text-black ${!open && 'rotate-180'}`}
-							onClick={() => (isMobile ? toggleMobileNav() : setOpen(!open))}
-						/>
-					</div>
+					{/* Replace with actual LogoIcon */}
+					<span className="text-lg font-bold">Logo</span>
+				</div>
 
-					{/* Body */}
-					<ul className="flex-1">
-						<div className="-mx-4 mb-4 h-[1px] bg-[#DEDEDE]"></div>
-						{menuItems.map((item, index) => {
-							const isActive = active === item.label
-							return (
-								<Popover key={index} label={item.label} show={!isMobile && !open}>
-									<li>
-										<Link
-											href={item.path}
-											onClick={() => setActive(item.label)}
-											className={`group flex items-center gap-2 rounded-lg p-3 ${
-												isActive ? 'bg-[#FFFFFF] text-black' : 'text-[#A0A2A0] hover:bg-[#E5E5E5]'
-											}`}
-										>
-											<div>{item.icon(isActive)}</div>
-											{(open || isMobile) && <p>{item.label}</p>}
-										</Link>
-									</li>
-								</Popover>
-							)
-						})}
-					</ul>
+				{/* Hamburger Menu - Always Visible */}
+				<button onClick={toggleSidebar} className="ml-auto rounded p-1 transition-all duration-200">
+					{/* MenuIcon rotates when sidebar is collapsed */}
+					<span
+						className={`h-6 w-6 cursor-pointer text-[#A0A2A0] transition-transform duration-100 hover:text-black ${
+							isExpanded ? 'rotate-0' : 'rotate-180'
+						}`}
+					>
+						☰
+					</span>
+				</button>
+			</div>
 
-					{/* Footer */}
-					<div className="mt-auto">
-						<div className="-mx-4 h-[1px] bg-[#DEDEDE]"></div>
-						<ul className="pb-4">
-							{menuItemsFooter.map((item, index) => {
-								const isActive = active === item.label
-								return (
-									<Popover key={index} label={item.label} show={!isMobile && !open}>
-										<li>
-											<Link
-												href={item.path}
-												onClick={() => setActive(item.label)}
-												className={`group flex items-center gap-2 rounded-lg p-3 ${
-													isActive ? 'bg-[#FFFFFF] text-black' : 'text-[#A0A2A0] hover:bg-[#E5E5E5]'
-												}`}
-											>
-												<div>{item.icon(isActive)}</div>
-												{(open || isMobile) && <p>{item.label}</p>}
-											</Link>
-										</li>
-									</Popover>
-								)
-							})}
-						</ul>
-					</div>
-				</nav>
-			)}
-		</section>
+			{/* Menu Items - Positioned at the Top */}
+			<ul className="space-y-1 p-2">
+				{menuItems.map((item) => (
+					<li key={item.path}>
+						<Link href={item.path}>
+							<div className="hover:bg-gray-200 group flex items-center rounded-md p-2">
+								{item.icon(isExpanded)}
+								<span
+									className={`ml-3 text-sm transition-opacity ${
+										isExpanded ? 'opacity-100' : 'hidden opacity-0'
+									}`}
+								>
+									{item.label}
+								</span>
+							</div>
+						</Link>
+					</li>
+				))}
+			</ul>
+
+			{/* Menu Items Footer - Positioned at the Bottom */}
+			<div className="absolute bottom-4 w-full px-2">
+				<ul className="space-y-1">
+					{menuItemsFooter.map((item) => (
+						<li key={item.path}>
+							<Link href={item.path}>
+								<div className="hover:bg-gray-200 group flex items-center rounded-md p-2">
+									{item.icon(isExpanded)}
+									<span
+										className={`ml-3 text-sm transition-opacity ${
+											isExpanded ? 'opacity-100' : 'hidden opacity-0'
+										}`}
+									>
+										{item.label}
+									</span>
+								</div>
+							</Link>
+						</li>
+					))}
+				</ul>
+			</div>
+		</nav>
 	)
 }
+
+export default SideNav
